refactor: move dcim types into 'types' pkg
