+++ dcim/console.go
@@ -11,56 +11,15 @@ import (
 	"github.com/neverbeencloser/gonautobot/types"
 )
 
-type (
-	// ConsolePort : Data representation of a Console Port in Nautobot.
-	ConsolePort struct {
-		ID           string                 `json:"id"`
-		URL          string                 `json:"url"`
-		CustomFields map[string]interface{} `json:"custom_fields"`
-		Device       struct {
-			ID          string `json:"id"`
-			URL         string `json:"url"`
-			Name        string `json:"name"`
-			DisplayName string `json:"display_name"`
-		} `json:"device"`
-		Name                  string            `json:"name"`
-		Label                 string            `json:"label"`
-		Description           string            `json:"description"`
-		ConnectedEndpointType string            `json:"connected_endpoint_type"`
-		ConnectedEndpoint     ConnectedEndpoint `json:"connected_endpoint"`
-		Cable                 struct {
-			ID    string `json:"id"`
-			URL   string `json:"url"`
-			Label string `json:"label"`
-		} `json:"cable"`
-		Tags []types.Tag `json:"tags"`
-	}
-
-	// ConnectedEndpoint : Struct representing the far-side of the console connection.
-	ConnectedEndpoint struct {
-		ID     string `json:"id"`
-		URL    string `json:"url"`
-		Device struct {
-			ID          string `json:"id"`
-			URL         string `json:"url"`
-			Name        string `json:"name"`
-			DisplayName string `json:"display_name"`
-		} `json:"device"`
-		Name  string `json:"name"`
-		Cable string `json:"cable"`
-		Port  int
-	}
-)
-
 // ConsoleConnectionFilter : Method used to fetch a list of console connections for a device.
-func (c *Client) ConsoleConnectionFilter(q *url.Values) ([]ConsolePort, error) {
+func (c *Client) ConsoleConnectionFilter(q *url.Values) ([]types.ConsolePort, error) {
 	req, err := c.Request(http.MethodGet, "dcim/console-connections/", nil, q)
 	if err != nil {
 		return nil, err
 	}
 
 	resp := new(types.ResponseList)
-	ret := make([]ConsolePort, 0)
+	ret := make([]types.ConsolePort, 0)
 	if err = c.UnmarshalDo(req, resp); err != nil {
 		return ret, err
 	}
