+++ dcim/interface.go
@@ -1,65 +1,24 @@
 package dcim
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/url"
 
 	"github.com/neverbeencloser/gonautobot/core"
 	"github.com/neverbeencloser/gonautobot/types"
-	"github.com/neverbeencloser/gonautobot/types/nested"
-)
-
-type (
-	// Interface : defines an interface as represented in Nautobot
-	//
-	// CablePeer and ConnectedEndpoint must be decoded dynamically; they are dependent
-	// on the value of 'CablePeerType' and 'ConnectedEndpointType', e.g., "dcim.interface"
-	Interface struct {
-		ID                         string                 `json:"id"`
-		Bridge                     *Interface             `json:"bridge"`
-		Cable                      *nested.Cable          `json:"cable"`
-		CablePeer                  json.RawMessage        `json:"cable_peer"`
-		CablePeerType              *string                `json:"cable_peer_type"`
-		ConnectedEndpoint          json.RawMessage        `json:"connected_endpoint"`
-		ConnectedEndpointReachable bool                   `json:"connected_endpoint_reachable"`
-		ConnectedEndpointType      *string                `json:"connected_endpoint_type"`
-		CountIPAddresses           int                    `json:"count_ipaddresses"`
-		Created                    string                 `json:"created"`
-		CustomFields               map[string]interface{} `json:"custom_fields"`
-		Description                string                 `json:"description"`
-		Device                     *Device                `json:"device"`
-		Display                    string                 `json:"display"`
-		Enabled                    bool                   `json:"enabled"`
-		Label                      string                 `json:"label"`
-		Lag                        *Interface             `json:"lag"`
-		LastUpdated                string                 `json:"last_updated"`
-		MACAddress                 *string                `json:"mac_address"`
-		MgmtOnly                   bool                   `json:"mgmt_only"`
-		Mode                       *types.LabelValue      `json:"mode"`
-		MTU                        *int                   `json:"mtu"`
-		Name                       string                 `json:"name"`
-		NotesURL                   string                 `json:"notes_url"`
-		ParentInterface            *Interface             `json:"parent_interface"`
-		TaggedVLANs                []nested.VLAN          `json:"tagged_vlans"`
-		Tags                       []types.Tag            `json:"tags"`
-		Type                       *types.LabelValue      `json:"type"`
-		UntaggedVLAN               *nested.VLAN           `json:"untagged_vlan"`
-		URL                        string                 `json:"url"`
-	}
 )
 
 // InterfaceGet : Go function to process requests for the endpoint: /api/dcim/interfaces/:id/
 //
 // https://demo.nautobot.com/api/docs/#/dcim/dcim_interfaces_retrieve
-func (c *Client) InterfaceGet(uuid string) (*Interface, error) {
+func (c *Client) InterfaceGet(uuid string) (*types.DeviceInterface, error) {
 	req, err := c.Request(http.MethodGet, fmt.Sprintf("dcim/interfaces/%s/", url.PathEscape(uuid)), nil, nil)
 	if err != nil {
 		return nil, err
 	}
 
-	ret := new(Interface)
+	ret := new(types.DeviceInterface)
 	err = c.UnmarshalDo(req, ret)
 	return ret, err
 }
@@ -67,9 +26,9 @@ func (c *Client) InterfaceGet(uuid string) (*Interface, error) {
 // InterfaceFilter : Go function to process requests for the endpoint: /api/dcim/interfaces/
 //
 // https://demo.nautobot.com/api/docs/#/dcim/dcim_interfaces_list
-func (c *Client) InterfaceFilter(q *url.Values) ([]Interface, error) {
-	resp := make([]Interface, 0)
-	if err := core.Paginate[Interface](c.Client, "dcim/interfaces/", q, &resp); err != nil {
+func (c *Client) InterfaceFilter(q *url.Values) ([]types.DeviceInterface, error) {
+	resp := make([]types.DeviceInterface, 0)
+	if err := core.Paginate[types.DeviceInterface](c.Client, "dcim/interfaces/", q, &resp); err != nil {
 		return nil, err
 	}
 	return resp, nil
