+++ tests/dcim_interface_test.go
@@ -1,11 +1,12 @@
 package nautobot_test
 
 import (
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"net/url"
 	"path"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"gopkg.in/h2non/gock.v1"
 )
