+++ types/dcim_devicefamily.go
@@ -7,15 +7,14 @@ import (
 )
 
 type (
-	// Namespace : Represents a namespace in Nautobot.
-	Namespace struct {
+	// DeviceFamily represents a device family in Nautobot's DCIM module.
+	DeviceFamily struct {
 		ID           uuid.UUID      `json:"id"`
 		Created      time.Time      `json:"created"`
 		CustomFields map[string]any `json:"custom_fields"`
 		Description  string         `json:"description"`
 		Display      string         `json:"display"`
 		LastUpdated  time.Time      `json:"last_updated"`
-		Location     *Location      `json:"location"`
 		Name         string         `json:"name"`
 		NaturalSlug  string         `json:"natural_slug"`
 		NotesURL     string         `json:"notes_url"`
@@ -24,13 +23,11 @@ type (
 		URL          string         `json:"url"`
 	}
 
-	// NewNamespace : Represents a new namespace to be created in Nautobot.
-	NewNamespace struct {
-		Name          string         `json:"name"`
-		CustomFields  map[string]any `json:"custom_fields,omitempty"`
-		Description   string         `json:"description,omitempty"`
-		Location      string         `json:"location,omitempty"`
-		Relationships map[string]any `json:"relationships,omitempty"`
-		Tags          []string       `json:"tags,omitempty"`
+	// NewDeviceFamily represents the data required to create a new device family.
+	NewDeviceFamily struct {
+		Name         string         `json:"name"`
+		CustomFields map[string]any `json:"custom_fields,omitempty"`
+		Description  string         `json:"description,omitempty"`
+		Tags         []string       `json:"tags,omitempty"`
 	}
 )
