+++ types/dcim_platform.go
@@ -1,0 +1,43 @@
+package types
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type (
+	// Platform represents a device platform in Nautobot.
+	Platform struct {
+		ID                    uuid.UUID      `json:"id"`
+		Created               time.Time      `json:"created"`
+		CustomFields          map[string]any `json:"custom_fields"`
+		Description           string         `json:"description"`
+		DeviceCount           int            `json:"device_count"`
+		Display               string         `json:"display"`
+		LastUpdated           time.Time      `json:"last_updated"`
+		Manufacturer          *Manufacturer  `json:"manufacturer"`
+		Name                  string         `json:"name"`
+		NapalmArgs            any            `json:"napalm_args"`
+		NapalmDriver          string         `json:"napalm_driver"`
+		NaturalSlug           string         `json:"natural_slug"`
+		NetworkDriver         string         `json:"network_driver"`
+		NetworkDriverMappings map[string]any `json:"network_driver_mappings"`
+		NotesURL              string         `json:"notes_url"`
+		ObjectType            string         `json:"object_type"`
+		URL                   string         `json:"url"`
+		VMCount               int            `json:"virtual_machine_count"`
+	}
+
+	// NewPlatform represents the data required to create a new platform.
+	NewPlatform struct {
+		Name          string         `json:"name"`
+		CustomFields  map[string]any `json:"custom_fields,omitempty"`
+		Description   string         `json:"description,omitempty"`
+		Manufacturer  string         `json:"manufacturer,omitempty"`
+		NapalmArgs    any            `json:"napalm_args,omitempty"`
+		NapalmDriver  string         `json:"napalm_driver,omitempty"`
+		NetworkDriver string         `json:"network_driver,omitempty"`
+		Relationships map[string]any `json:"relationships,omitempty"`
+	}
+)
