+++ types/dcim_rackgroup.go
@@ -7,30 +7,32 @@ import (
 )
 
 type (
-	// Namespace : Represents a namespace in Nautobot.
-	Namespace struct {
+	// RackGroup : Data type entry for a RackGroup in Nautobot.
+	RackGroup struct {
 		ID           uuid.UUID      `json:"id"`
 		Created      time.Time      `json:"created"`
 		CustomFields map[string]any `json:"custom_fields"`
 		Description  string         `json:"description"`
 		Display      string         `json:"display"`
 		LastUpdated  time.Time      `json:"last_updated"`
-		Location     *Location      `json:"location"`
+		Location     Location       `json:"location"`
 		Name         string         `json:"name"`
 		NaturalSlug  string         `json:"natural_slug"`
 		NotesURL     string         `json:"notes_url"`
 		ObjectType   string         `json:"object_type"`
-		Tags         []Tag          `json:"tags"`
+		Parent       *RackGroup     `json:"parent"`
+		RackCount    int            `json:"rack_count"`
+		TreeDepth    *int           `json:"tree_depth"`
 		URL          string         `json:"url"`
 	}
 
-	// NewNamespace : Represents a new namespace to be created in Nautobot.
-	NewNamespace struct {
-		Name          string         `json:"name"`
+	// NewRackGroup represents the structure for creating a new RackGroup in Nautobot.
+	NewRackGroup struct {
 		CustomFields  map[string]any `json:"custom_fields,omitempty"`
 		Description   string         `json:"description,omitempty"`
-		Location      string         `json:"location,omitempty"`
+		Location      string         `json:"location"`
+		Name          string         `json:"name"`
+		Parent        string         `json:"parent,omitempty"`
 		Relationships map[string]any `json:"relationships,omitempty"`
-		Tags          []string       `json:"tags,omitempty"`
 	}
 )
