+++ types/virtualization_cluster.go
@@ -2,8 +2,6 @@ package types
 
 import (
 	"time"
-
-	"github.com/neverbeencloser/gonautobot/types/nested"
 )
 
 // Cluster : Cerebro Cluster data representation in Nautobot.
@@ -20,11 +18,11 @@ type Cluster struct {
 		ObjectType string `json:"object_type"`
 		URL        string `json:"url"`
 	} `json:"cluster_type"`
-	ClusterGroup any              `json:"cluster_group"`
-	Tenant       *Tenant          `json:"tenant"`
-	Location     *nested.Location `json:"location"`
-	Created      time.Time        `json:"created"`
-	LastUpdated  time.Time        `json:"last_updated"`
-	NotesURL     string           `json:"notes_url"`
-	CustomFields map[string]any   `json:"custom_fields"`
+	ClusterGroup any            `json:"cluster_group"`
+	Tenant       *Tenant        `json:"tenant"`
+	Location     *Location      `json:"location"`
+	Created      time.Time      `json:"created"`
+	LastUpdated  time.Time      `json:"last_updated"`
+	NotesURL     string         `json:"notes_url"`
+	CustomFields map[string]any `json:"custom_fields"`
 }
